+++ samila/functions.py
@@ -13,9 +13,12 @@ from .params import DEFAULT_BACKGROUND_COLOR, DEFAULT_SPOT_SIZE, DEFAULT_PROJECT
 from .params import Projection, VALID_COLORS, HEX_COLOR_PATTERN, NFT_STORAGE_API, NFT_STORAGE_LINK, OVERVIEW
 from .params import DATA_TYPE_ERROR, CONFIG_TYPE_ERROR, PLOT_DATA_ERROR, CONFIG_NO_STR_FUNCTION_ERROR
 from .params import NO_FIG_ERROR_MESSAGE, FIG_SAVE_SUCCESS_MESSAGE, NFT_STORAGE_SUCCESS_MESSAGE, SAVE_NO_DATA_ERROR
+from .params import INVALID_COLOR_TYPE_ERROR
+from .params import BOTH_COLOR_COMPLEMENT_WARNING, COLOR_NOT_FOUND_WARNING
 from .params import DATA_SAVE_SUCCESS_MESSAGE, SEED_LOWER_BOUND, SEED_UPPER_BOUND
 from .params import ELEMENTS_LIST, ARGUMENTS_LIST, OPERATORS_LIST, RANDOM_COEF_LIST
 from .errors import samilaDataError, samilaPlotError, samilaConfigError
+from warnings import warn
 
 
 def random_equation_gen():
@@ -95,26 +98,114 @@ def distance_calc(s1, s2):
     return distances[-1]
 
 
-def filter_color(color):
+def is_valid_color(color):
     """
-    Filter given color and return it.
+    Check that input color format is valid or not.
 
     :param color: given color
-    :type color: str or tuple
-    :return: filtered version of color
+    :type color: any format
+    :return: result as bool
     """
-    if isinstance(color, tuple):
-        return color
+    if color == None:
+        return True
+    try:
+        _ = matplotlib.colors.to_hex(color)
+        return True
+    except ValueError:
+        return False
+
+
+def color_complement(color):
+    """
+    Calculate complement color.
+
+    :param color: given color (hex format)
+    :type color: str
+    :return: complement color (hex format) as str
+    """
+    color = color[1:]
+    color = int(color, 16)
+    comp_color = 0xFFFFFF ^ color
+    comp_color = "#%06x" % comp_color
+    return comp_color
+
+
+def filter_color(color, bgcolor):
+    """
+    Filter given color and bgcolor.
+
+    :param color: given color
+    :type color: any format
+    :param bgcolor: given background color
+    :type bgcolor: any format
+    :return: filtered version of color and bgcolor
+    """
+    color = select_color(color)
+    bgcolor = select_color(bgcolor)
+    if color == "COMPLEMENT" and bgcolor == "COMPLEMENT":
+        warn(BOTH_COLOR_COMPLEMENT_WARNING, RuntimeWarning)
+        return None, None
+    if color == "COMPLEMENT":
+        bgcolor = matplotlib.colors.to_hex(bgcolor)
+        color = color_complement(bgcolor)
+    if bgcolor == "COMPLEMENT":
+        color = matplotlib.colors.to_hex(color)
+        bgcolor = color_complement(color)
+    return color, bgcolor
+
+
+def select_color(color):
+    """
+    Select color and return it.
+
+    :param color: given color
+    :type color: any format
+    :return: color
+    """
     if isinstance(color, str):
+        if color.upper() == "TRANSPARENT":
+            return "TRANSPARENT"
+        if color.upper() == "COMPLEMENT":
+            return "COMPLEMENT"
         if color.upper() == "RANDOM":
             return random_hex_color_gen()
         if re.match(HEX_COLOR_PATTERN, color):
-            return color
+            return color.lower()
         distance_list = list(map(lambda x: distance_calc(color, x),
                                  VALID_COLORS))
         min_distance = min(distance_list)
-        return VALID_COLORS[distance_list.index(min_distance)]
-    return None
+        most_similar_color = VALID_COLORS[distance_list.index(min_distance)]
+        if min_distance != 0:
+            warn(
+                COLOR_NOT_FOUND_WARNING.format(
+                    color,
+                    most_similar_color),
+                RuntimeWarning)
+        return most_similar_color
+    if is_valid_color(color):
+        return color
+    raise samilaPlotError(INVALID_COLOR_TYPE_ERROR)
+
+
+def set_background(bgcolor, fig, ax):
+    """
+    Set background for figure and axis.
+
+    :param bgcolor: given background color
+    :type bgcolor: any format
+    :param fig: figure
+    :type fig: matplotlib.figure.Figure
+    :param ax: axis
+    :type ax: matplotlib.axes._subplots.AxesSubplot
+    :return: None
+    """
+    if bgcolor == "TRANSPARENT":
+        ax.patch.set_visible(False)
+        fig.patch.set_visible(False)
+        return
+    fig.set_facecolor(bgcolor)
+    ax.set_facecolor(bgcolor)
+    return
 
 
 def filter_projection(projection):
@@ -196,7 +287,7 @@ def plot_params_filter(
         raise samilaPlotError(PLOT_DATA_ERROR.format(1))
     if g.data2 is None:
         raise samilaPlotError(PLOT_DATA_ERROR.format(2))
-    color, bgcolor = map(filter_color, [color, bgcolor])
+    color, bgcolor = filter_color(color, bgcolor)
     projection = filter_projection(projection)
     alpha = filter_float(alpha)
     linewidth = filter_float(linewidth)